src_python/requirements.txt
discord
openai
python-dotenv
pandas
pydantic[email]